// src/domain/types.ts
export type EffectDto =
| { kind: "HCDistortion"; data: HcDistortionDto }
| { kind: "SCDistortion"; data: ScDistortionDto }
| { kind: "Delay"; data: DelayDto }
| { kind: "Cabinet"; data: CabinetDto }
| { kind: "Wah"; data: WahDto }
;

export interface TunerContractDto {
  live_tuner_event: string;
}

export interface AudioSettingsDto {
  input_device_name: string;
  output_device_name: string;
  input_sample_rate: number;
  output_sample_rate: number;
  input_channels: number;
  output_channels: number;
  audio_driver: string;
}

export interface AlgorithmicLatencyDto {
  processor_name: string;
  latency_samples: number;
  latency_ms: number;
}

export interface IrProfileDto {
  file_name: string;
  label: string;
  is_custom: boolean;
  is_in_use: boolean;
}

export interface ToneStackDto {
  bass: number;
  middle: number;
  treble: number;
}

export interface SpectrumSnapshotDto {
  sample_rate_hz: number;
  frequencies_hz: number[];
  magnitudes: number[];
  level_db: number;
}

export interface MidiMappingDto {
  cc_number: number;
  channel: number;
  effect_id: string;
  parameter: MidiTargetParameter;
}


export type MidiTargetParameter = "ToggleBypass" | "WahPedalPosition" | "DelayTime" | "DelayLevel" | "DistortionLevel" | "DistortionThreshold";

export interface AmpConfigDto {
  master_volume: number;
  is_active: boolean;
  channels: ChannelDto[];
  current_channel: string;
  audio_settings: AudioSettingsDto;
  midi_bindings: MidiMappingDto[];
}
// ...
export interface ChannelDto {
  id: string;
  name: string;
  gain: number;
  tone_stack: ToneStackDto;
  volume: number;
  effect_chain: EffectDto[];
}